travel-agent1/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
langchain>=0.1.0
langchain-openai
langchain-community
tavily-python
